fixes/1200933785056563 - Remove Key word from resolvers collections and contracts
