+++ src/LDL/Framework/Base/Collection/Resolver/Contracts/FullResolverInterface.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Framework\Base\Collection\Resolver\Contracts;
+
+interface FullResolverInterface extends CustomResolverInterface, DuplicateResolverInterface, NullResolverInterface
+{
+
+}
