+++ src/LDL/Framework/Base/Collection/Resolver/Key/StringKeyResolver.php
@@ -1,0 +1,62 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Framework\Base\Collection\Resolver\Key;
+
+use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
+use LDL\Framework\Base\Collection\Resolver\Contracts\FullResolverInterface;
+
+class StringKeyResolver implements FullResolverInterface
+{
+    /**
+     * @var string
+     */
+    private $prefix;
+
+    /**
+     * @var int
+     */
+    private $limit;
+
+    public function __construct(string $prefix='_', int $limit=100000)
+    {
+        $this->prefix = $prefix;
+        $this->limit = $limit;
+    }
+
+    public function resolveCustom(CollectionInterface $collection, $key, $item = null, ...$params)
+    {
+        return sprintf(
+            '%s%s%s',
+            $key ?? 'key',
+            $this->prefix,
+            $collection->count()
+        );
+    }
+
+    public function resolveNull(CollectionInterface $collection, $item=null, ...$params)
+    {
+        return $this->resolveCustom($collection, null, $item, ...$params);
+    }
+
+    public function resolveDuplicate(CollectionInterface $collection, $key, $item=null, ...$params)
+    {
+        if(!is_string($key)) {
+            return $key;
+        }
+
+        if(!$collection->hasKey($key)){
+            return $key;
+        }
+
+        $i = 0;
+
+        while($i++ < $this->limit){
+            $newKey = "$key{$this->prefix}$i";
+            if(!$collection->hasKey($newKey)){
+                return $newKey;
+            }
+        }
+
+        throw new \LogicException("Could not resolve a suitable key for $key");
+    }
+}
